changed midterm
added array ellement that allows you to click and add new baqlls to the game. also added a grey scale background to the mix to make it harder to see the ball. i am interested in changing the game from a "two player game" to a one player game by adding gravity and also switching the direction of the game field.