Gun prototype created
not sure if i will keep it how it is or change it in the future. but the gun can now rotate and shows up on the screen. next step is to incorperate the key comands to allow you to move around the screen while shooting. next i will start creating the bullets