Created Restraints + movement
The tak can now move up, down, left, and right using w,s,a,d keys. Was unable to make the key release function properly so that the tank would not be constantly moving but it kept freezing up, so instead I just decided to remove it all together. Also created restraints so the tank will now stay in the confinement of the screen.

Still working on getting the bullets to function